backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openai
python-multipart